Use fault injection patch for subtask processor
