+++ mars/deploy/oscar/tests/fault_injection_config.yml
@@ -1,5 +1,3 @@
 "@inherits": '@default'
 third_party_modules:
   - mars.services.tests.fault_injection_patch
-subtask:
-  subtask_processor_cls: mars.services.subtask.worker.tests.FaultInjectionSubtaskProcessor
