+++ mars/services/subtask/worker/tests/__init__.py
@@ -12,5 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .subtask_processor import CheckedSubtaskProcessor, \
-    FaultInjectionSubtaskProcessor
+from .subtask_processor import CheckedSubtaskProcessor
+
